RoseModel/requirements.txt
fastapi
uvicorn
httpx
chromadb
pydantic
sse-starlette
pyyaml
litellm
jinja2
python-multipart
